requirements.txt
flask==2.3.3
flask-socketio==5.3.0
flask-admin==1.6.1
flask-sqlalchemy==3.1.1
eventlet==0.33.3
requests==2.31.0
paho-mqtt==1.6.1
